mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "async>=0.6.2",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.14",
    "langchain-pinecone>=0.2.5",
    "mcp[cli]>=1.6.0",
    "pinecone>=6.0.2",
    "pinecone-notebooks>=0.1.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
